Video write bug fix.

+++ yolox/utils/visualize.py
@@ -11,7 +11,7 @@ __all__ = ["vis"]
 
 def vis(img, boxes, scores, cls_ids, conf=0.5, class_names=None):
     if isinstance(img, torch.Tensor):
-        img = img.cpu().float().numpy()
+        img = img.cpu().numpy()
 
     for i in range(len(boxes)):
         box = boxes[i]
